Changed flight pattern back to SF
Flight pattern may be swapped out at any time for a completely differrent pattern of BaseFlightPattern type. Therefore, I want to keep this serializedfield I think so a designer can change it in teh editor as needed... RigidBody won't change during play or need a designer to have more than 1, so I left that as a regular field.